Fixed newlib spu scripting error.

+++ buildit.sh
@@ -212,9 +212,11 @@ buildnewlib() {
 			--prefix=$PS3DEV/$FOLDER && \
 		$MAKE $MAKEOPTS && \
 		$MAKE install
-		cp -r $PS3DEV/$FOLDER/$NEWLIBTARGET/lib $PS3DEV/$FOLDER/$TARGET/
-		cp -r $PS3DEV/$FOLDER/$NEWLIBTARGET/include $PS3DEV/$FOLDER/$TARGET/
-		rm -rf $PS3DEV/$FOLDER/$NEWLIBTARGET
+		if [ "$TARGET" != "$NEWLIBTARGET" ]; then
+			cp -r $PS3DEV/$FOLDER/$NEWLIBTARGET/lib $PS3DEV/$FOLDER/$TARGET/
+			cp -r $PS3DEV/$FOLDER/$NEWLIBTARGET/include $PS3DEV/$FOLDER/$TARGET/
+			rm -rf $PS3DEV/$FOLDER/$NEWLIBTARGET
+		fi
 	) || die "Error building newlib for target $TARGET"
 }
